requirements.txt
flask
flask-smorest
python-dotenv

sqlalchemy
flask-sqlalchemy

flask-migrate

gunicorn
psycopg2-binary

celery
redis
